+++ index.html
@@ -1,4 +1,5 @@
 <!DOCTYPE html>
+<html lang="en">
 <head>
     <title>Комментарий</title>
     <meta charset="utf-8"/>
@@ -28,5 +29,5 @@
 </div>
 </body>
 
-<script type="module" src="maim.js"></script>
+<script type="module" src="index.js"></script>
 </html>
