--- maim.js
@@ -1,108 +1,0 @@
-'use strict'
-
-const nameInput = document.querySelector('.add-form-name')
-const textInput = document.querySelector('.add-form-text')
-const button = document.querySelector('.add-form-button')
-const commentsList = document.querySelector('.comments')
-
-const comments = [
-    {
-        name: 'Глеб Фокин',
-        date: new Date(2022, 1, 12, 12, 18),
-        text: 'Это будет первый комментарий на этой странице',
-        likes: 3,
-        isLiked: false,
-    },
-    {
-        name: 'Варвара Н.',
-        date: new Date(2022, 1, 13, 19, 22),
-        text: 'Мне нравится как оформлена эта страница! ❤',
-        likes: 75,
-        isLiked: false,
-    },
-]
-
-function renderComments(comments) {
-    commentsList.innerHTML = comments
-        .map(
-            (comment, index) => `
-     <li class="comment" data-index="${index}" >
-        <div class="comment-header">
-            <div>${comment.name}</div>
-            <div>${comment.date.toLocaleString('ru', {
-                year: 'numeric',
-                month: 'long',
-                day: 'numeric',
-                hour: 'numeric',
-                minute: 'numeric',
-            })}</div>
-            </div>
-     <div class="comment-body">
-     <div class="comment-text">${comment.text}</div>
-     </div>
-           <div class="comment-footer">
-             <div class="likes">
-               <span class="likes-counter">${comment.likes}</span>
-               <button class="like-button ${comment.isLiked ? '-active-like' : ''}" data-index="${index}"></button>
-             </div>
-           </div>
-     </li>
-    `,
-        )
-        .join('')
-
-    document.querySelectorAll('.like-button').forEach((button) => {
-        button.addEventListener('click', (event) => {
-            event.stopPropagation()
-            const index = event.target.dataset.index
-            const comment = comments[index]
-
-            if (comment.isLiked) {
-                comment.likes--
-                comment.isLiked = false
-            } else {
-                comment.likes++
-                comment.isLiked = true
-            }
-            renderComments(comments)
-        })
-    })
-
-    document.querySelectorAll('.comment').forEach((commentElement) => {
-        commentElement.addEventListener('click', (event) => {
-            if (!event.target.closest('.like-button')) {
-                const index = commentElement.dataset.index
-                const comment = comments[index]
-                textInput.value = `${comment.name}, : ${comment.text}`
-                textInput.focus()
-            }
-        })
-    })
-}
-
-renderComments(comments)
-
-button.addEventListener('click', () => {
-    const name = nameInput.value.replaceAll('<', '&lt;').replaceAll('>', '&gt;')
-    const text = textInput.value.replaceAll('<', '&lt;').replaceAll('>', '&gt;')
-
-    if (!name || !text) {
-        alert('Пожалуйста, заполните все поля')
-        return
-    }
-
-    const newComment = {
-        name: name,
-        date: new Date(),
-        text: text,
-        likes: 0,
-        isLiked: false,
-    }
-    comments.push(newComment)
-
-    renderComments(comments)
-
-    nameInput.value = ''
-    textInput.value = ''
-})
-console.log('It works!')
