+++ modules/comments.js
@@ -1,0 +1,16 @@
+export const comments = [
+    {
+        name: 'Глеб Фокин',
+        date: new Date(2022, 1, 12, 12, 18),
+        text: 'Это будет первый комментарий на этой странице',
+        likes: 3,
+        isLiked: false,
+    },
+    {
+        name: 'Варвара Н.',
+        date: new Date(2022, 1, 13, 19, 22),
+        text: 'Мне нравится как оформлена эта страница! ❤',
+        likes: 75,
+        isLiked: false,
+    },
+]
